Refs #2233 #3148 #3150 #3215 #3152 Added tests.
Refs #3233. Added Output Best Effort test.

Refs #3148. Addapt interaction and cross_serialization tests. Removed xrce_serialization test.

Refs #3148. Added Gateway.

Refs #3148. Client interaction test strucure base finished.

Refs #3150, #3150, #2315, #3152. Finish tests. The test with lost have errors.

Removed unused files and declarations

Added examples with cxx compatibility.

Addapted client.h to nuttx.

Fixed defined error.

Fix Bugs at tests and modify logs.

Refs #3150. Fix test bugs.

Ref #3150. Fix input bug at lost messages.

Remove a warning from example code.

+++ examples/PublishHelloWorld/HelloWorldWriter.c
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/*! 
+/*!
  * @file HelloWorldWriter.c
  * This source file contains the write function of the described types in the IDL file.
  *
@@ -29,15 +29,15 @@ bool mr_write_HelloWorld_topic(mrSession* session, mrStreamId stream_id, mrObjec
 {
     bool written = false;
 
-    uint16_t topic_length = (uint16_t)HelloWorld_size_of_topic(topic, 0);
-    uint16_t payload_length = 0;
+    uint32_t topic_length = HelloWorld_size_of_topic(topic, 0);
+    uint32_t payload_length = 0;
     payload_length += 4; //request_id + object_id
     payload_length += 4; //topic_length (remove in future version to be compliant)
 
     MicroBuffer mb;
     if(prepare_stream_to_write(&session->streams, stream_id, payload_length + topic_length + SUBHEADER_SIZE, &mb))
     {
-        (void) write_submessage_header(&mb, SUBMESSAGE_ID_WRITE_DATA, payload_length + topic_length, FORMAT_DATA);
+        (void) write_submessage_header(&mb, SUBMESSAGE_ID_WRITE_DATA, (uint16_t)(payload_length + topic_length), FORMAT_DATA);
 
         WRITE_DATA_Payload_Data payload;
         init_base_object_request(session, datawriter_id, &payload.base);
